@@ -10,13 +10,16 @@
     }
   ],
   "require": {
-    "php": ">=7.1",
+    "php": ">=7.4",
     "ext-mbstring": "*"
   },
   "autoload": {
     "psr-4": {
       "Ueef\\Sanitizer\\": "src"
     }
+  },
+  "require-dev": {
+    "phpunit/phpunit": "^8.5"
   },
   "autoload-dev": {
     "psr-4": {
